Added exception handling

+++ todo-app/src/main/java/com/harmbrugge/todographql/TodoController.java
@@ -1,5 +1,6 @@
 package com.harmbrugge.todographql;
 
+import com.harmbrugge.todographql.exception.TodoNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
@@ -7,6 +8,8 @@ import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
 
 import java.time.LocalDateTime;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Controller
 public class TodoController {
@@ -29,20 +32,30 @@ public class TodoController {
 
     @MutationMapping
     public Todo updateTodo(@Argument Integer id, @Argument String info, @Argument String name, @Argument boolean done) {
-        Todo todo = todoRepository.findById(id).get();
-        todo.setName(name);
-        todo.setInfo(info);
-        todo.setId(id);
-        todo.setDone(done);
-        todo.setUpdatedAt(LocalDateTime.now());
-        return todoRepository.save(todo);
+        Optional<Todo> todoDatabase = todoRepository.findById(id);
+        if (todoDatabase.isPresent()) {
+            Todo todo = todoDatabase.get();
+            todo.setName(name);
+            todo.setInfo(info);
+            todo.setId(id);
+            todo.setDone(done);
+            todo.setUpdatedAt(LocalDateTime.now());
+            return todoRepository.save(todo);
+        } else {
+            throw new TodoNotFoundException(String.format("No todo item with id %d", id));
+        }
     }
 
     @MutationMapping
     public Todo deleteTodo(@Argument Integer id) {
-        Todo todo = todoRepository.findById(id).get();
-        todoRepository.deleteById(id);
-        return todo;
+        Optional<Todo> todoDatabase = todoRepository.findById(id);
+        if (todoDatabase.isPresent()) {
+            todoRepository.deleteById(id);
+            return todoDatabase.get();
+        }
+        else {
+            throw new TodoNotFoundException(String.format("No todo item with id %d", id));
+        }
     }
 
 }
